requirements.txt
aiokafka==0.12.0
amqp==5.3.1
annotated-types==0.7.0
async-timeout==5.0.1
asyncpg==0.30.0
billiard==4.2.2
celery==5.3.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dill==0.4.0
dotenv==0.9.9
greenlet==3.2.4
inquirerpy==0.3.4
kombu==5.5.4
markdown-it-py==4.0.0
mdurl==0.1.2
multiprocess==0.70.14
orjson==3.11.0
packaging==25.0
palitra==0.0.1.post1
pfzy==0.3.4
pillow==12.0.0
prompt_toolkit==3.0.52
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==5.0.1
rich==14.2.0
robyn==0.72.2
rustimport==1.3.4
six==1.17.0
SQLAlchemy==2.0.44
toml==0.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvloop==0.21.0
vine==5.1.0
watchdog==4.0.1
wcwidth==0.2.14
